% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaAPI.R
\name{beaAPI}
\alias{beaAPI}
\alias{beaChangeDates}
\alias{beaDFtoXTS}
\title{API Bureau of Economic Analysis}
\usage{
beaAPI(api.param = stop("'api.param' must be specified"), curl = NULL,
  query = FALSE, raw = FALSE)

beaDFtoXTS(data = stop("'data' must be provided"))

beaChangeDates(str)
}
\arguments{
\item{api.param}{a list of parameters specific to the API}

\item{curl}{optional, \code{CURL} handle created with \code{RCurl::getCurlHandle()}}

\item{query}{the API query}

\item{raw}{return unparsed API response}

\item{data}{a data frame created with \code{beaJSONtoDF}}

\item{str}{a character string with BEA dates, e.g. \code{"2000"}, \code{"2000Q1"}, \code{"2000M01"}}
}
\value{
The function creates an URL with with the specified
parameters, retrieves the JSON string and transforms to R list
}
\description{
Connect to BEA datasets
}
\details{
These functions allow connecting to BEA time series in JSON format with a valid user ID obtained from http://www.bea.gov/API/signup/index.cfm
}
\examples{
\dontrun{
curl <- RCurl::getCurlHandle()
api.param <- list(USERID = "api.key",
                  METHOD = "GETDATASETLIST",
                  RESULTFORMAT = "JSON")
List <- beaAPI(api.param = api.param, curl = curl)
df <- beaJSONtoDF(List = List, third = 1)
}
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{JSON}

