% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsoApiGPG.R
\name{nsoApiGPG}
\alias{nsoApiGPG}
\title{Source apiKey}
\usage{
nsoApiGPG(file = system.file("apiKey.R.gpg", package = "nsoApi"),
  gpg = file.path("D:", "GPG", "gpg2.exe"), passphrase = NULL,
  shell = file.path("C:", "Windows", "System32", "cmd.exe"), keep = FALSE)
}
\arguments{
\item{file}{path to gpg key file}

\item{gpg}{path to GnuPG executable}

\item{passphrase}{character password used for decryption. suggested to store as environment variable and use with \code{Sys.getenv()}}

\item{shell}{if on Windows, you may need to specify the path to cmd.exe}

\item{keep}{boolean set to TRUE if you want to keep the decoded file}
}
\value{
Creates decrypted file in the package installation folder. The key file is gpg encryped. In Windows, use GPG Plugin Portable: http://portableapps.com/apps/security/gpg-plugin-portable
}
\description{
Source apiKey list PGP encrypted R script
}
\details{
Source apiKey list PGP encrypted R script with keys for various webservices of National Statics Offices.
}
\examples{
\dontrun{
nsoApiGPG(keep = TRUE)
file.apiKey <- system.file("apiKey.R", package = "nsoApi")

if (file.exists(file.apiKey)) {
  filecon <- file(file.apiKey)
  cat(paste0(readLines(con = filecon), '\\n'))
  close(filecon)
}
}
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{GPG}

