% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onsAPI.R
\name{onsAPI}
\alias{onsAPI}
\alias{onsCsvData}
\alias{onsDFgather}
\title{API ONS}
\usage{
onsAPI(method = "collections", context = "Economic",
  api.key = stop("'api.key' must be provided"))

onsCsvData(api.param = stop("'api.param' must be provided"),
  dataset = stop("'dataset' must be provided"), query = FALSE,
  curl = NULL, ...)

onsDFgather(data = stop("'data' must be provided"))
}
\arguments{
\item{method}{}

\item{context}{}

\item{api.key}{}

\item{api.param}{a list with parameters used to construct the query, see examples.}

\item{dataset}{the dataset to download.}

\item{query}{logical, return the query containing the download link.}

\item{curl}{optional, \code{CURL} handle created with \code{RCurl::getCurlHandle()}.}

\item{data}{a data frame created with \code{onsAPI}}
}
\value{
The main function creates an URL with with the specified parameters, retrieves the zip file, extracts a csv file that is read into R and returned as data frame. Additional functions convert the returned data frame to an xts object.
}
\description{
Information retrieval from ONS webservices
}
\details{
Retrieve information from ONS API in CSV format.
}
\examples{
onsAPI(method = "contexts", api.key = api.key)
onsAPI(method = "collections", context = "Economic", api.key = api.key)
## Within a group there is a series for each combination of dimension items in the segment. In a time series dataset there will be several observations within a series, but in the case of Census data there is only one time (Census Day 2011).
api.param <- list(context = "Census",
                  geog = "2011WARDH",
                  totals = "false",
                  apikey = "xPuqnMzZ01")
data.QS208EW <- onsAPI(api.param = api.param, dataset = "QS208EW")
data.QS104EW <- onsAPI(api.param = api.param, dataset = "QS104EW")
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{csv}

