% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pxwebAPI.R
\name{pxwebAPI}
\alias{pxwebAPI}
\alias{pxwebChangeDates}
\alias{pxwebDFtoXTS}
\title{API PX-Web}
\usage{
pxwebDFtoXTS(data = stop("'data' must be provided"))

pxwebChangeDates(str)
}
\arguments{
\item{data}{a data frame created with \code{get_pxweb_data}}

\item{str}{a character string with PXWEB dates, e.g. \code{"2000"}, \code{"2000K1"}, \code{"2000M01" }}
}
\value{
Convert data frame retrieved by \code{get_pxweb_data} into xts object.
}
\description{
Information retrieval from PX-Web webservices
}
\details{
Retrieve information from PX-Web data APIs.
}
\examples{
dims_list <- list(SNI2007 = c('*'),
                  Transaktionspost = c('*'),
                  ContentsCode = c('*'),
                  Tid = c('*'))
req.uri <- "http://api.scb.se/OV0104/v1/doris/en/ssd"
data.df <- pxweb::get_pxweb_data(url = req.uri, dims = dims_list, clean = TRUE)

## convert to xts object, e.g. for use in \\code{dygraphs} package
data.xts <- pxwebDFtoXTS(data = data.df)
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{PC-Axis}

