% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sdmxRead}
\alias{sdmxRead}
\title{Query Data OECD.Stat}
\usage{
sdmxRead(api = "http://stats.oecd.org/SDMX-JSON", scheme = "data",
  DSD = "BTDIXE_I4", filter = list(COU = c("ESP", "DEU"), FLW = c("IMPO"),
  PAR = c("WOR"), EUC = c("TOTAL", "INT"), IND = c("DTOTAL", "D01T03"), VAL =
  c("VALUE")), append = paste0("/all?", paste("json-lang=en", "detail=Full",
  "dimensionAtObservation=AllDimensions", "startPeriod=2000", "endPeriod=2010",
  sep = "&")), query = FALSE)
}
\arguments{
\item{api}{an API address}

\item{scheme}{an API scheme. Available schemes \code{data}, \code{codelist}.}

\item{DSD}{a datastructure definition identified by the triplet \code{[collection; country; indicator]}.}

\item{filter}{for scheme "data": a named list of filters passed to the API. The position of list items corresponds to the API filter dimensions. Each list item is either empty (no filter on dimension) or a character vector containing dimension members to be included in the results. Dimension members can be obtained from \code{scheme="codelist"} and a codelist item, e.g. "CL_ECO_ISIC4".}

\item{append}{append string to the dimension url.}

\item{query}{logical to return SDMX http url only.}
}
\description{
Query data from OECD.Stat using SDMX
}
\details{
Helper function to efficiently query data from OECD.Stat Extracts using SDMX-JSON API.
}
\examples{
## list dimension members
test.codelist <- sdmxRead(DSD = "BTDIXE_I4", scheme = "codelist")
## retrieve data
filter.list <- list(COU = c("ESP", "DEU"),
                    FLW = c("IMPO"),
                    PAR = c("WOR"),
                    EUC = c("TOTAL", "INT"),
                    IND = c("DTOTAL", "D01T03"),
                    VAL = c("VALUE"))
url.append <- paste0("/all?", paste("json-lang=en", "detail=Full", "dimensionAtObservation=AllDimensions", "startPeriod=1990", "endPeriod=2000", sep = "&"))
test.data <- sdmxRead(DSD = "BTDIXE_I4", filter = filter.list)
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\seealso{
https://data.oecd.org/api/sdmx-json-documentation, http://cran.r-project.org/web/packages/jsonlite/index.html
}
\keyword{OECD.Stat,}
\keyword{SDMX}

