% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectInputs}
\alias{selectInputs}
\title{Shiny Extra}
\usage{
selectInputs(data = stop("'data' must be provided"), names = NULL,
  prefix = "", ...)
}
\arguments{
\item{data}{a data frame}

\item{names}{a character vectors with columns of data.}

\item{prefix}{a character string to prepend inputIds.}
}
\description{
Additions to RStudio shiny functions
}
\details{
Minor additions to existing RStudio Shiny functions.
}
\examples{
cat(selectInputs(data = data, names = c("fruits", "color"), prefix = "dim_")
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\seealso{
\code{https://github.com/rstudio/shiny}
}
\keyword{shiny}

