% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssbAPI.R
\name{ssbAPI}
\alias{ssbAPI}
\title{API SSB}
\usage{
ssbAPI(dataset, curl = NULL)
}
\value{
A data frame converted from JSON-Stat format
}
\description{
Information retrieval from Statistics Norway API
}
\details{
Retrieve information from Statistics Norway API in \code{JSON-Stat} format.
}
\examples{
req.uri <- "http://data.ssb.no/api/v0/dataset/44631.json?lang=en"
data.df <- rjstat::fromJSONstat(x = req.uri)

## convert to xts object, e.g. for use in \\code{dygraphs} package
data.xts <- pxwebDFtoXTS(data = data.df)
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{JSON-Stat}

