% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdsAPI.R
\name{wdsAPI}
\alias{wdsAPI}
\alias{wdsDFtoXTS}
\alias{wdsChangeDates}
\title{API WDS}
\usage{
wdsAPI(dataset = stop("'dataset' must be provided"), query = FALSE,
  curl = NULL, ...)

wdsDFtoXTS(data = stop("'data' must be provided"))

wdsChangeDates(str)
}
\arguments{
\item{data}{a data frame created with \code{wdsAPI}}

\item{str}{a character string with WDS dates, e.g. \code{"2001"}}
}
\value{
The \code{wdsAPI} function creates an URL with with the specified parameters, retrieves the zip file, extracts a csv file that is read into R and returned as data frame. \code{wdsDFtoXTS} converts the returned data frame to an xts object.
}
\description{
Information retrieval from WDS webservices
}
\details{
Retrieve information from WDS data API in csv format.
}
\examples{
\dontrun{
data.df <- wdsAPI(dataset = "2810027")

## convert to xts object, e.g. for use in \\code{dygraphs} package
data.xts <- wdsDFtoXTS(data = data.df)
}
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{SOAP}
