% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beaAPI}
\alias{beaAPI}
\alias{beaJSONtoDF}
\title{API Bureau of Economic Analysis}
\usage{
beaAPI(api.param = stop("'api.param' must be specified"),
  curl = getCurlHandle(), query = FALSE, raw = FALSE)

beaJSONtoDF(List = stop("'List' must be specified"), third = 1)
}
\arguments{
\item{api.param}{}

\item{curl}{}

\item{query}{}

\item{raw}{}

\item{List}{}

\item{third}{}
}
\value{
The function creates an URL with with the specified
parameters, retrieves the JSON string and transforms to R list
}
\description{
Connect to BEA datasets
}
\details{
These functions allow connecting to BEA time series in JSON format
with a valid user ID obtained from
\link{http://www.bea.gov/API/signup/index.cfm}
}
\examples{
api.param <- list(USERID = api.key,
                  METHOD = "GETDATASETLIST",
                  RESULTFORMAT = "JSON")
beaAPI(api.param = api.param, curl = curl)
}

