% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beaAPI}
\alias{beaAPI}
\alias{beaChangeDates}
\title{API Bureau of Economic Analysis}
\usage{
beaAPI(api.param = stop("'api.param' must be specified"), curl = NULL,
  query = FALSE, raw = FALSE)

beaChangeDates(str)
}
\arguments{
\item{api.param}{}

\item{curl}{optional, \code{CURL} handle created with \code{RCurl::getCurlHandle()}}

\item{query}{}

\item{raw}{}

\item{str}{a character string with BEA dates, e.g. \code{"2000"}, \code{"2000Q1"}, \code{"2000M01" }}
}
\value{
The function creates an URL with with the specified
parameters, retrieves the JSON string and transforms to R list
}
\description{
Connect to BEA datasets
}
\details{
These functions allow connecting to BEA time series in JSON format
with a valid user ID obtained from
\link{http://www.bea.gov/API/signup/index.cfm}
}
\examples{
api.key <- "7023E825-15FF-488D-B8D9-D70E6F67D439"
curl <- RCurl::getCurlHandle()
api.param <- list(USERID = api.key,
                  METHOD = "GETDATASETLIST",
                  RESULTFORMAT = "JSON")
List <- beaAPI(api.param = api.param, curl = curl)
df <- beaJSONtoDF(List = List, third = 1)
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{JSON}

