% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsODataAPI.R
\name{cbsODataAPI}
\alias{cbsODataAPI}
\alias{cbsOdataDFgather}
\alias{cbsOdataDFtoXTS}
\alias{cbsODataTables}
\alias{cbsODataChangeDates}
\title{Query CBS OData API}
\usage{
cbsODataAPI(api = stop("'api' must be provided"),
  DSD = stop("'DSD' must be provided"), scheme = NULL, filter = NULL,
  append = NULL, query = FALSE, curl = NULL)

cbsOdataDFgather(data = stop("'data' must be provided"))

cbsOdataDFtoXTS(data = stop("'data' must be provided"))

cbsODataTables(url = "https://opendata.cbs.nl/ODataCatalog/Tables",
  fields = c("Identifier", "Title", "Frequency", "Period"))

cbsODataChangeDates(str)
}
\arguments{
\item{api}{an API address}

\item{DSD}{a datastructure definition identified by the triplet \code{[collection; country; indicator]}.}

\item{scheme}{an API scheme. Available scheme "data".}

\item{filter}{for scheme "data": a named list of filters passed to the API. The position of list items corresponds to the API filter dimensions. Each list item is either empty (no filter on dimension) or a character vector containing dimension members to be included in the results. Dimension members can be obtained from \code{scheme="codelist"} and a codelist item, e.g. "CL_ECO_ISIC4".}

\item{append}{append string to the dimension url.}

\item{query}{logical to return OData http url only.}

\item{curl}{optional, \code{CURL} handle created with \code{RCurl::getCurlHandle()}}

\item{data}{a data frame created with \code{cbsODataAPI}}

\item{url}{location of table list XML document.}

\item{fields}{character vector to extract content properties of entries.}

\item{str}{a character string with CBS OData dates, e.g. \code{"1995JJ00"}}

\item{xml.list}{a character string returned from}
}
\description{
Retrieve information from CBS Open Data API
}
\details{
Retrieve information from CBS Netherlands Statline Open Data API using JSON format.
}
\examples{
\dontrun{
api <- "https://opendata.cbs.nl/ODataApi/OData/"
DSD <- "82572ENG" # Input-Output: "83068ENG"

cbsODataAPI(api=api, DSD=DSD, scheme = NULL)
str(cbsODataAPI(api=api, DSD=DSD, scheme="SectorBranchesSIC2008", query=FALSE))
str(cbsODataAPI(api=api, DSD=DSD, scheme="TypedDataSet"))
}
}
\seealso{
\code{https://github.com/object/Simple.OData.Client}
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\keyword{JSON}
\keyword{OData}
