% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInputs.R
\name{selectInputs}
\alias{selectInputs}
\title{Multiple selectInput}
\usage{
selectInputs(list = stop("'list' must be provided"), prefix = "",
  minsize = 4, ...)
}
\arguments{
\item{list}{a named list object.}

\item{prefix}{a character string to prepend inputIds.}
}
\description{
Create multiple shiny \code{selectInput} widgets
}
\details{
Create list of shiny ui widgets based on list with label and choices.
}
\examples{
\dontrun{
listexample <- list(fruits = c("mango", "apple"),
                    color = c("orange", "green"))

selectInputs(list = listexample, prefix = "dim_")

## generate list from data frame
df <- data.frame(fruits = c("mango", "apple", "mango"),
                 color = c("orange", "green", "red"))

list.df <- apply(df, 2, unique)

selectInputs(list = list.df, prefix = "dim_")
}
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\seealso{
\code{https://github.com/rstudio/shiny}
}
\keyword{shiny}

