% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statatAPI.R
\name{statatAPI}
\alias{statatAPI}
\alias{statatChangeDates}
\alias{statatDFtoXTS}
\title{Query Statistik Austria open.data}
\usage{
statatAPI(dataset = stop("'DSD' must be provided"), query = FALSE,
  curl = NULL)

statatDFtoXTS(data = stop("'data' must be provided"))

statatChangeDates(str)
}
\arguments{
\item{query}{logical to return OData http url only.}

\item{curl}{optional, \code{CURL} handle created with \code{RCurl::getCurlHandle()}}

\item{data}{a data frame created with \code{get_pxweb_data}}

\item{str}{a character string with Statistik Austria open.data dates, e.g. \code{"A10-1976"}, \code{"197311"}, \code{"VPIZR-201101" }}

\item{DSD}{a dataset identifier.}
}
\description{
Retrieve information from Statistik Austria open.data
}
\details{
Retrieve information from Statistik Austria open.data using CSV format.
}
\examples{
dataset <- "OGD_vgr001_VGRJahresR_1"
statatAPI(dataset = dataset)
}
\author{
Bo Werth <bo.werth@gmail.com>
}
\seealso{
\code{http://http://data.statistik.gv.at/web}
}
\keyword{JSON}

